Implemented authentication middleware and first access configuration

+++ src/features/user/service.ts
@@ -1,13 +1,17 @@
 import type { ID } from "@shared/types/id.js";
 import { EntityNotFoundError } from "@shared/errors.js";
 
-import type { NewUser, User } from "@features/user/entity.js";
+import type { PasswordHasher } from "@features/authentication/password-hasher.js";
+import type { LoggedUser, NewUser, User } from "@features/user/entity.js";
 import type { UserRepository } from "@features/user/repository.js";
+import { UnauthorizedAccessError } from "@features/authentication/exceptions.js";
+import { UserRole } from "@features/user/enums.js";
 import { DuplicateEmailError } from "@features/user/exceptions.js";
 
 export class UserService {
     constructor(
-        private repository: UserRepository
+        private repository: UserRepository,
+        private passwordHasher: PasswordHasher,
     ) { }
 
     public async getAll() {
@@ -26,14 +30,24 @@ export class UserService {
         return this.repository.findByEmail(email);
     }
 
-    public async create(data: NewUser) {
+    public async create(loggedUser: LoggedUser, data: NewUser) {
+        if (loggedUser.role !== UserRole.ADMIN) {
+            throw new UnauthorizedAccessError("Only admin users can create new users.");
+        }
+
         await this.ensureUsernameIsUnique(data.username);
         await this.ensureEmailIsUnique(data.email);
+
+        data.password = await this.passwordHasher.hash(data.password);
 
         return this.repository.create(data);
     }
 
-    public async update(id: ID, data: User) {
+    public async update(loggedUser: LoggedUser, id: ID, data: User) {
+        if (loggedUser.role !== UserRole.ADMIN && loggedUser.id !== id) {
+            throw new UnauthorizedAccessError("It is not allowed to update other users' information.");
+        }
+
         await this.ensureUsernameIsUnique(data.username);
         await this.ensureEmailIsUnique(data.email);
 
@@ -49,7 +63,11 @@ export class UserService {
         return this.repository.update(id, existingUser);
     }
 
-    public async delete(id: string) {
+    public async delete(loggedUser: LoggedUser, id: string) {
+        if (loggedUser.role !== UserRole.ADMIN && loggedUser.id !== id) {
+            throw new UnauthorizedAccessError("It is not allowed to update other users' information.");
+        }
+
         return this.repository.delete(id);
     }
